add NioUtil
