+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * 【poi    】     Excel07SaxReader拆分出SheetDataSaxHandler
 * 【core   】     CollUtil.addAll增加判空（pr#228@Gitee）
 * 【core   】     修正DateUtil.betweenXXX注释错误（issue#I28XGW@Gitee）
+* 【core   】     增加NioUtil
 
 ### Bug修复
 * 【cache  】     修复Cache中get重复misCount计数问题（issue#1281@Github）
@@ -29,6 +30,7 @@
 * 【db     】     修复表名包含点导致的问题（issue#1300@Github）
 * 【poi    】     修复xdr:row标签导致的问题（issue#1297@Github）
 * 【core   】     修复FileUtil.loopFiles使用FileFilter无效问题（issue#I28V48@Gitee）
+* 【extra  】     修复JschUtil.execByShell返回空的问题（issue#1067@Github）
 
 -------------------------------------------------------------------------------------------------------------
